requirements.txt
streamlit>=1.37
pandas>=2.1
numpy>=1.26
requests>=2.31
yfinance>=0.2.52
scipy>=1.11
python-dateutil>=2.9
pytz>=2024.1
